Add scale mode to Texture
Feature doesn't work too well as etting the scale mode for 1 texture
will affect all the other textures generated by the same file (like for
sprites).

Leaving it as a "known bug" for now as it's not a big deal and there are
workarounds like separating textures based on scale mode.